Scoremaker: WIP, moving file handling logic to ever_given
